requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.9.0
httpx==0.27.2
livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.3
livekit-plugins-noise-cancellation~=0.2
livekit-api>=1.0.7,<2
livekit-plugins-bey~=1.3
python-dateutil==2.9.0.post0
